refactor: extracted zod configuration
